Rename ShadowSpringBootConditionTest
