+++ shardingsphere-features/shardingsphere-shadow/shardingsphere-shadow-spring/shardingsphere-shadow-spring-boot-starter/src/test/java/org/apache/shardingsphere/shadow/spring/boot/ShadowSpringBootConditionTest.java
@@ -17,44 +17,32 @@
 
 package org.apache.shardingsphere.shadow.spring.boot;
 
-import org.apache.shardingsphere.shadow.spring.boot.condition.ShadowSpringBootCondition;
+import org.apache.shardingsphere.shadow.api.config.ShadowRuleConfiguration;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.junit.MockitoJUnitRunner;
-import org.springframework.boot.autoconfigure.condition.ConditionOutcome;
-import org.springframework.context.annotation.ConditionContext;
-import org.springframework.core.type.AnnotatedTypeMetadata;
-import org.springframework.mock.env.MockEnvironment;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
+import javax.annotation.Resource;
+import java.util.Collections;
 
-@RunWith(MockitoJUnitRunner.class)
-public final class ShadowSpringBootConditionTest {
+import static org.hamcrest.CoreMatchers.is;
+import static org.junit.Assert.assertThat;
+
+@RunWith(SpringJUnit4ClassRunner.class)
+@SpringBootTest(classes = ShadowSpringBootConditionTest.class)
+@SpringBootApplication
+@ActiveProfiles("shadow")
+public class ShadowSpringBootConditionTest {
+    
+    @Resource
+    private ShadowRuleConfiguration shadowRuleConfiguration;
     
     @Test
-    public void assertNotMatch() {
-        MockEnvironment mockEnvironment = new MockEnvironment();
-        mockEnvironment.setProperty("spring.shardingsphere.rules.encrypt.encryptors.aes_encryptor.type", "AES");
-        ConditionContext context = mock(ConditionContext.class);
-        AnnotatedTypeMetadata metadata = mock(AnnotatedTypeMetadata.class);
-        when(context.getEnvironment()).thenReturn(mockEnvironment);
-        ShadowSpringBootCondition condition = new ShadowSpringBootCondition();
-        ConditionOutcome matchOutcome = condition.getMatchOutcome(context, metadata);
-        assertFalse(matchOutcome.isMatch());
-    }
-    
-    @Test
-    public void assertMatch() {
-        MockEnvironment mockEnvironment = new MockEnvironment();
-        mockEnvironment.setProperty("spring.shardingsphere.rules.shadow.column", "user_id");
-        ConditionContext context = mock(ConditionContext.class);
-        AnnotatedTypeMetadata metadata = mock(AnnotatedTypeMetadata.class);
-        when(context.getEnvironment()).thenReturn(mockEnvironment);
-        ShadowSpringBootCondition condition = new ShadowSpringBootCondition();
-        ConditionOutcome matchOutcome = condition.getMatchOutcome(context, metadata);
-        assertTrue(matchOutcome.isMatch());
+    public void assertAlgorithmProviderBean() {
+        assertThat(shadowRuleConfiguration.getColumn(), is("shadow"));
+        assertThat(shadowRuleConfiguration.getShadowMappings(), is(Collections.singletonMap("ds", "shadow_ds")));
     }
 }
