+++ shardingsphere-features/shardingsphere-shadow/shardingsphere-shadow-spring/shardingsphere-shadow-spring-boot-starter/src/test/resources/application-shadow.properties
@@ -17,4 +17,3 @@
 
 spring.shardingsphere.rules.shadow.column=shadow
 spring.shardingsphere.rules.shadow.shadowMappings.ds=shadow_ds
-
